examproject/requirements.txt
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
enum34==1.1.10
filelock==3.0.12
idna==3.2
peewee==3.14.4
platformdirs==2.2.0
psycopg2==2.9.1
python-decouple==3.4
PyYAML==5.4.1
requests==2.26.0
scv==2.0.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
virtualenv==20.7.2